+++ CHANGELOG.md
@@ -5,10 +5,20 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ### Added
-- Demo functionality for the generation of first image [#6](https://github.com/grhvaring/RTXOn/pull/6)
+- Read 3D scene to render from input file
+
+## [0.3.0] - 2022-05-25
+
+### Added
+- Path Tracing algorithm [#8](https://github.com/grhvaring/RTXOn/pull/8)
+
+## [0.2.0] - 2022-05-18
+
+### Added
+- Generation of a demo image [#6](https://github.com/grhvaring/RTXOn/pull/6)
 
 ### Fixed
-- Fix an issue with the vertical order of the images [#4](https://github.com/grhvaring/RTXOn/pull/5)
+- Images flipped upside down [#4](https://github.com/grhvaring/RTXOn/pull/5)
 
 ## [0.1.0] - 2022-04-27